Te_Po/services/ruru/ruru.py
# -*- coding: utf-8 -*-
# backend/services/ruru/main.py

from fastapi import FastAPI, UploadFile, File
import pytesseract
from PIL import Image
import openai
import io
from supabase import create_client, Client  # Assuming supabase-py package

# Set up FastAPI app
app = FastAPI()

# Supabase client setup (pseudo-code)
supabase_url = "your-supabase-url"
supabase_key = "your-anon-key"
supabase: Client = create_client(supabase_url, supabase_key)

@app.get("/status")
async def status():
    return {"status": "Ruru is awake", "language": "bilingual"}

@app.post("/summarise")
async def summarise(file: UploadFile = File(...)):
    # Read file content
    content = await file.read()
    # OCR and processing based on file type would go here
    # For example, using pytesseract if OCR is needed

    # Generate embeddings and summaries
    # Example: embeddings = generate_embeddings(content)
    # summaries = generate_summary(content)

    # Write results to Supabase
    # response = supabase.table("ruru_summaries").insert({...}).execute()

    return {"message": "File processed and summarised"}

if __name__ == "__main__":
    import uvicorn
    uvicorn.run(app, host="0.0.0.0", port=8004)