Te_Po/services/ruru/requirements.txt
# backend/services/ruru/requirements.txt

fastapi
uvicorn
openai
supabase
pytesseract
Pillow